+++ test/Feature/HtmlRawOutputSpec.hs
@@ -11,7 +11,7 @@ import Protolude  hiding (get)
 import SpecHelper (acceptHdrs)
 
 spec :: SpecWith Application
-spec = describe "When raw-output-media-types is set to \"text/html\"" $
+spec = describe "When raw-media-types is set to \"text/html\"" $
   it "can get raw output with Accept: text/html" $
            request methodGet "/rpc/welcome.html" (acceptHdrs "text/html") ""
              `shouldRespondWith`